+++ escort_flow_gurobi_v8.py
@@ -40,7 +40,7 @@ class SolverConfig:
     fractional_horizon: int
     integer_horizon: int
     epoch: int
-    time_limit: int
+    time_limit: float
     num_threads: int
     mip_focus: int
     warmstart_mode: str
@@ -351,20 +351,20 @@ class RollingHorizonGurobiSolver:
                 if move != stay_move
             ]
             for t in range(T + 1):
-                # (7) Escort conflict avoidance.
+                # (6) Escort conflict avoidance.
                 model.addConstr(
                     gp.quicksum(x_e[(move, t)] for move in self.network["cell_cover"][loc]) <= 1,
                     name=f"cell_cover_{loc[0]}_{loc[1]}_t{t}",
                 )
                 if not self.config.bnc:
-                    # (9) A target load must move if crossed by an escort.
+                    # (8) A target load must move if crossed by an escort.
                     model.addConstr(
                         1 - x_a[(stay_move, t)] >= gp.quicksum(
                             x_e[(move, t)] for move in self.network["cell_cover"][loc]
                         ),
                         name=f"stay_cover_{loc[0]}_{loc[1]}_t{t}",
                     )
-                    # (8) A target load cannot move unless crossed by an escort.
+                    # (7) A target load cannot move unless crossed by an escort.
                     for move in nonstay_target_moves:
                         model.addConstr(
                             x_a[(move, t)] <= gp.quicksum(
@@ -376,7 +376,7 @@ class RollingHorizonGurobiSolver:
 
         retrieve_total_constraint = None
         if self.config.full:
-            # (10) Every target load must eventually arrive at an output cell.
+            # (9) Every target load must eventually arrive at an output cell.
             arrival_moves_to_outputs = [
                 move for move in self.network["moves_a"]
                 if (move[2], move[3]) in self.output_set and (move[0], move[1]) != (move[2], move[3])
@@ -391,7 +391,7 @@ class RollingHorizonGurobiSolver:
                     move for move in self.network["incoming_a"][output]
                     if (move[0], move[1]) != (move[2], move[3])
                 ]
-                # (15) q stores the summed arrival times at each output cell.
+                # (14) q stores the summed arrival times at each output cell.
                 model.addConstr(
                     gp.quicksum(
                         (t + 1) * x_a[(move, t)]
@@ -403,7 +403,7 @@ class RollingHorizonGurobiSolver:
         bnc_stay_specs_by_t = None
         bnc_move_specs_by_t = None
         if self.config.bnc:
-            # Prepare separated representations of (8) and (9) for the callback.
+            # Prepare separated representations of (7) and (8) for the callback.
             bnc_stay_specs_by_t = {t: [] for t in range(T + 1)}
             bnc_move_specs_by_t = {t: [] for t in range(T + 1)}
             for loc in self.network["locations"]:
